backend/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
SQLAlchemy==2.0.15
django-cors-headers==4.0.0
pandas==2.0.1